gas_chem - using a separate file for generated code.
