+++ src/mam4xx/mam4.hpp
@@ -24,6 +24,7 @@
 #include <mam4xx/nucleate_ice.hpp>
 #include <mam4xx/nucleation.hpp>
 #include <mam4xx/rename.hpp>
+#include <mam4xx/generated_gas_chem.hpp>
 
 namespace mam4 {
